Add payment integration, delivery tracking, and improve error handling
- Restaurant: Add BigDecimal pricing, MenuCategory enum, timestamps, and enhanced error handling
- Order: Add payment processing (PaymentStatus, PaymentMethod), delivery fields, and customer APIs
- Notification: Add NotificationService with message templates
- Dispatch: Add race condition protection and improved logging
- All services: Clean up POMs and update configurations

+++ order-service/src/main/java/com/treatz/orderservice/dto/MenuItemResponseDTO.java
@@ -2,12 +2,14 @@ package com.treatz.orderservice.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class MenuItemResponseDTO {
     private Long id;
     private String name;
     private String description;
-    private Double price;
+    private BigDecimal price; // Updated to BigDecimal for proper money handling
     private boolean available;
     private String category;
     private String imageUrl;
