+++ restaurant-service/src/main/java/com/treatz/restaurantservice/dto/MenuItemResponseDTO.java
@@ -2,13 +2,15 @@ package com.treatz.restaurantservice.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class MenuItemResponseDTO {
     private Long id;
     private String name;
     private String description;
-    private Double price;
+    private BigDecimal price;
     private boolean available;
-    private String category;
+    private String category; // Keeping as String for JSON compatibility
     private String imageUrl;
 }
